Refactored the app and client fixtures to conftest.py
