+++ tests/test_views.py
@@ -1,32 +1,6 @@
 """Test the major views using flask's test_client()."""
 import pytest
-from run import create_app, db
-from flask_migrate import Migrate
-
-
-@pytest.fixture
-def app():
-    """Set up and tear down the test app."""
-    app = create_app('testing')
-    Migrate(app, db)
-
-    app.config['TESTING'] = True
-
-    app_context = app.app_context()
-    app_context.push()
-    db.create_all()
-
-    yield app
-
-    db.session.remove()
-    db.drop_all()
-    app_context.pop()
-
-
-@pytest.fixture
-def client(app):
-    """Set up and tear down the test client."""
-    yield app.test_client()
+from run import db
 
 
 @pytest.mark.parametrize("route", [
@@ -49,11 +23,12 @@ def test_404(client, route):
     assert "404" in response.status
 
 
-@pytest.mark.xfail
+@pytest.mark.xfail(strict=True)
 def test_500(client):
     """Check the error handler for 500 errors."""
-    response = client.put('/foo')
-    assert b'Page not found' in response.data
+    # Find a way to force or mock a server error.
+    response = client.get('/foo')
+    assert b'Internal Server Error' in response.data
     assert "500" in response.status
 
 
@@ -64,29 +39,29 @@ def test_dropped_db(client):
     assert "404" in client.get('/poem/Flea').status
 
 
-def test_reset_route(client):
+def test_reset_route(client_empty_db):
     """Check the /reset route."""
     # The database has tables in it
     assert "poems" in db.engine.table_names()
     # The database has no poems in it
-    assert b'Sorry' in client.get('/').data
-    assert "404" in client.get('/poem/Flea').status
+    assert b'Sorry' in client_empty_db.get('/').data
+    assert "404" in client_empty_db.get('/poem/Flea').status
 
     # /reset route injects sample poems into database
-    client.get('/reset')
-    assert b'Flea' in client.get('/').data
-    response = client.get('/poem/Flea')
+    client_empty_db.get('/reset')
+    assert b'Flea' in client_empty_db.get('/').data
+    response = client_empty_db.get('/poem/Flea')
     assert b'Flea' in response.data
     assert "200" in response.status
-    assert "404" in client.get('/poem/foo').status
+    assert "404" in client_empty_db.get('/poem/foo').status
 
     # Makes sure doing it twice doesn't break anything.
-    client.get('/reset')
-    assert b'Flea' in client.get('/').data
-    response = client.get('/poem/Flea')
+    client_empty_db.get('/reset')
+    assert b'Flea' in client_empty_db.get('/').data
+    response = client_empty_db.get('/poem/Flea')
     assert b'Flea' in response.data
     assert "200" in response.status
-    assert "404" in client.get('/poem/foo').status
+    assert "404" in client_empty_db.get('/poem/foo').status
 
 
 def test_about_default(client, monkeypatch):
